backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
